script: keygen.py: Change from `ecdsa` to `cryptography`
Changes the library used by `kegen.py` from `ecdsa` to `cryptography.
The reason for this change is `ecdsa` statement about itself:

"This library was not designed with security in mind. If you are
processing data that needs to be protected we suggest you use a
quality wrapper around OpenSSL. pyca/cryptography is one example of
such a wrapper. The primary use-case of this library is as a portable
library for interoperability testing and as a teaching tool."

Keygen is now changed to `cryptography` which uses OpenSSL as a backend.

Ref. NCSDK-18419

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>