caf: modules: power_manager: use sys_poweroff
Instead of pm_state_force, now unavailable to perform a system off. Note
that selecting POWEROFF in Kconfig is technically wrong. However, this
is a non-portable application (nRF specific) where POWEROFF will always
be selectable (except nrf53 cpunet, not a target). Fixing this properly
would require a long chain of changes, so I think it is better to allow
this select here.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>