samples: wifi: twt: Add traffic gen module
This module can be used to send traffic by establishing a control
connection based on TCP and then sending data traffic either TCP or UDP
and printing results at the end (both client and server).

This gives us explicit control over sending traffic rather than using
tools like Zperf, the primary user for this will be TWT sample, so, its
under the sample, but if we find more users then this can be moved to a
library.

For now only TCP data traffic support is added as its the primary
usecase.

Signed-off-by: Chiranjeevi Srikakulapu <chiranjeevi.srikakulapu@nordicsemi.no>