applications: nrf5340_audio: Error -22 on reboots
Occasionally during reboots, there will be an error
-22 as headset and/or gateway initializes. When this
happens on the Gateway, an <err> print will be generated.
A headset usually reports a warning in the same situation
due to PACs.

This has, for instance, been observed in a unicast
test-case with one gateway and one headset, where both
are halted and rebooted (using pinreset) simultaneously
multiple times.

This would appear to be that the gateway is trying to
stop a stream in a NULL unicast group (due to the reset).
It should be possible to store the state of the stream
state and only stop if playing (quick and tested to work
on this branch). Alternatively, do not perform any media
commands until the stream is configured, but what to do
with them is interesting.

Signed-off-by: Graham Wacey <graham.wacey@nordicsemi.no>